WIP: parent_id added to comments table

+++ app/Livewire/CommentItem.php
@@ -11,6 +11,14 @@ class CommentItem extends Component
 
     public bool $editing = false;
 
+    public bool $replying = false;
+
+    protected $listeners = [
+        'cancelEditing' => 'cancelEditing',
+        'commentUpdated' => 'commentUpdated',
+        'commentCreated' => 'commentCreated',
+    ];
+
     public function mount(Comment $comment)
     {
         $this->comment = $comment;
@@ -23,7 +31,17 @@ class CommentItem extends Component
 
     public function deleteComment()
     {
+        $user = auth()->user();
+        if (!$user) {
+            return $this->redirect('/login');
+        }
+
+        if ($this->comment->user_id != $user->id) {
+            return response('You are not allowed to perform this action', 403);
+        }
+
         $id = $this->comment->id;
+
         $this->comment->delete();
         $this->emitUp('commentDeleted', $id);
     }
@@ -31,5 +49,26 @@ class CommentItem extends Component
     public function startCommentEdit()
     {
         $this->editing = true;
+    }
+
+    public function cancelEditing()
+    {
+        $this->editing = false;
+        $this->replying = false;
+    }
+
+    public function commentUpdated()
+    {
+        $this->editing = false;
+    }
+
+    public function startReply()
+    {
+        $this->replying = true;
+    }
+
+    public function commentCreated()
+    {
+        $this->replying = false;
     }
 }
